cms-agent/requirements.txt
Here are the contents for the file /cms-agent/cms-agent/requirements.txt:

google-adk
flask
requests
pandas
numpy